+++ llvm/docs/ReleaseNotes.rst
@@ -49,6 +49,11 @@ Non-comprehensive list of changes in this release
   the name used in the `def X : Target` definition to the call to
   `RegisterTarget`.
 
+* The ``Debugify`` pass was added to ``opt`` to facilitate testing of debug
+  info preservation. This pass attaches synthetic ``DILocations`` and
+  ``DIVariables`` to the instructions in a ``Module``. The ``CheckDebugify``
+  pass determines how much of the metadata is lost.
+
 * Note..
 
 .. NOTE